+++ lib/features/Restaurant_Menu/model/dish.dart
@@ -3,10 +3,10 @@ class DishData {
   String? dish_name;
   String? dish_description;
   String? dish_imageurl;
-  int? dish_price;
+  double? dish_price;
   String? dish_schedule_meal;
   String? cusine;
-  int? dish_discount;
+  double? dish_discount;
   String? category_id;
   bool? isVariation;
 
@@ -44,8 +44,8 @@ class DishData {
     return DishData(
         dishid: json['id'] ?? 0,
         dish_name: json['dish_name'] ?? '',
-        dish_price: json['dish_price'] ?? 0,
-        dish_discount: json['dish_discount'] ?? 0,
+        dish_price: json['dish_price'] ?? 0.0,
+        dish_discount: json['dish_discount'] ?? 0.0,
         dish_description: json['dish_description'] ?? '',
         dish_imageurl: json['dish_imageurl'] ?? '',
         dish_schedule_meal: json['dish_schedule_meal'] ?? '',
